+++ packages/messaging/src/jobs/event/queue.ts
@@ -1,3 +1,4 @@
 export enum EventsQueue {
   EVENTS = 'event-queue',
+  FALLBACK_EVENTS = 'fallback-event-queue',
 }