+++ packages/messaging/src/jobs/user/queue.ts
@@ -1,3 +1,4 @@
 export enum UserQueue {
   USER = 'user-queue',
+  FALLBACK_USER = 'fallback-user-queue',
 }