+++ packages/post-processors/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.3.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.6.0
+  - @volontariapp/shared@0.6.0
+  - @volontariapp/database@3.2.13
+  - @volontariapp/outbox@0.9.17
+
 ## 1.3.16
 
 ### Patch Changes